Download geospatial data from GISCO API - Eurostat

giscoR is a API package that helps to retrieve data from Eurostat - GISCO (the Geographic Information System of the COmmission)

Source

GISCO webpage

Details

giscoR package

Package:	giscoR

Type:	Package

Version:	See sessionInfo() or DESCRIPTION file

Date:	2020

License:	GPL-3

LazyLoad:	yes

Note

COPYRIGHT NOTICE

When data downloaded from this page is used in any printed or electronic publication, in addition to any other provisions applicable to the whole Eurostat website, data source will have to be acknowledged in the legend of the map and in the introductory page of the publication with the following copyright notice:

EN: (C) EuroGeographics for the administrative boundaries

FR: (C) EuroGeographics pour les limites administratives

DE: (C) EuroGeographics bezüglich der Verwaltungsgrenzen

For publications in languages other than English, French or German, the translation of the copyright notice in the language of the publication shall be used.

If you intend to use the data commercially, please contact EuroGeographics for information regarding their licence agreements.

References

See citation("giscoR")

Author

dieghernan, https://github.com/dieghernan/

Examples

library(giscoR)