Attribution when publishing GISCO data

Get the legal text to be used along with the data downloaded with this package

gisco_attributions(lang = "en", copyright = FALSE)

Arguments

lang
Language (two-letter ISO code). See https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes and Details.

copyright
Boolean. Whether to display the copyright notice or not on the console.

Value

A string with the attribution to be used.

Details

Current languages supported are:

"en" - English

"da" - Danish

"de" - German

"es" - Spanish

"fi" - Finish

"fr" - French

"no" - Norwegian

"sv" - Swedish

Consider contributing if you spot any mistake or want to add a new language.

Note

COPYRIGHT NOTICE

When data downloaded from GISCO is used in any printed or electronic publication, in addition to any other provisions applicable to the whole Eurostat website, data source will have to be acknowledged in the legend of the map and in the introductory page of the publication with the following copyright notice:

EN: (C) EuroGeographics for the administrative boundaries

FR: (C) EuroGeographics pour les limites administratives

DE: (C) EuroGeographics bezuglich der Verwaltungsgrenzen

For publications in languages other than English, French or German, the translation of the copyright notice in the language of the publication shall be used.

If you intend to use the data commercially, please contact EuroGeographics for information regarding their licence agreements.

Examples

en <- gisco_attributions()

gisco_attributions(lang = "es", copyright = TRUE)

#> 
#>     COPYRIGHT NOTICE
#> 
#>     When data downloaded from GISCO
#>     is used in any printed or electronic publication,
#>     in addition to any other provisions applicable to
#>     the whole Eurostat website, data source will have
#>     to be acknowledged in the legend of the map and in
#>     the introductory page of the publication with the
#>     following copyright notice:
#> 
#>     - EN: (C) EuroGeographics for the administrative boundaries
#>     - FR: (C) EuroGeographics pour les limites administratives
#>     - DE: (C) EuroGeographics bezuglich der Verwaltungsgrenzen
#> 
#>     For publications in languages other than English,
#>     French or German, the translation of the copyright
#>     notice in the language of the publication shall be
#>     used.
#> 
#>     If you intend to use the data commercially, please
#>     contact EuroGeographics for information regarding
#>     their licence agreements.
#> 
#>       

#> [1] "© Eurogeographics para los límites administrativos"

gisco_attributions(lang = "XXX")

#> [1] "Language not supported, switching to English. Consider contributing"

#> [1] "© EuroGeographics for the administrative boundaries"