Bulk download from GISCO API

Downloads zipped data from GISCO and extract them on the cache_dir folder.

Usage

gisco_bulk_download(
  id_giscoR = c("countries", "coastallines", "communes", "lau", "nuts", "urban_audit"),
  year = "2016",
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE,
  resolution = "10",
  ext = c("geojson", "shp", "svg", "json", "gdb"),
  recursive = TRUE
)

Source

https://gisco-services.ec.europa.eu/distribution/v2/

Arguments

id_giscoR

Type of dataset to be downloaded. Values supported are:

"coastallines".

"communes".

"countries".

"lau".

"nuts".

"urban_audit".

year

Release year of the file. See Details.

cache_dir

A path to a cache directory. See About caching.

update_cache

A logical whether to update cache. Default is FALSE. When set to TRUE it would force a fresh download of the source .geojson file.

verbose

Logical, displays information. Useful for debugging, default is FALSE.

resolution

Resolution of the geospatial data. One of

"60": 1:60million

"20": 1:20million

"10": 1:10million

"03": 1:3million

"01": 1:1million

ext

Extension of the file(s) to be downloaded. Formats available are "geojson", "shp", "svg", "json", "gdb". See Details.

recursive

Tries to unzip recursively the zip files (if any) included in the initial bulk download (case of ext = "shp").

Value

Silent function.

Details

See the years available in the corresponding functions:

gisco_get_coastallines().

gisco_get_communes().

gisco_get_countries().

gisco_get_lau().

gisco_get_nuts().

gisco_get_urban_audit().

The usual extension used across giscoR is "geojson", however other formats are already available on GISCO.

About caching

You can set your cache_dir with gisco_set_cache_dir().

Sometimes cached files may be corrupt. On that case, try re-downloading the data setting update_cache = TRUE.

If you experience any problem on download, try to download the corresponding .geojson file by any other method and save it on your cache_dir. Use the option verbose = TRUE for debugging the API query.

For a complete list of files available check gisco_db.

See also

Examples

# \dontrun{

# Countries 2016 - It would take some time
gisco_bulk_download(id_giscoR = "countries", resolution = "60")
# }