Bulk download from GISCO API

Downloads zipped data from GISCO and extract them on the cache_dir folder.

gisco_bulk_download(
  id_giscoR = "countries",
  year = "2016",
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE,
  resolution = "10",
  ext = "geojson",
  recursive = TRUE
)

Arguments

id_giscoR
Type of dataset to be downloaded. Values supported are:

"coastallines"

"communes"

"countries"

"lau"

"nuts"

"urban_audit"

year
Release year. See Details.

cache_dir
a path to a cache directory. The directory have to exist. The NULL (default) uses and creates /gisco directory in the temporary directory from tempdir. The directory can also be set with options(gisco_cache_dir = "path/to/dir").

update_cache
a logical whether to update cache. Default is FALSE. When set to TRUE it would force a fresh download of the source .geojson file.

verbose
Display information. Useful for debugging, default if FALSE.

resolution
Resolution of the geospatial data. One of

"60" (1:60million),

"20" (1:20million)

"10" (1:10million)

"03" (1:3million) or

"01" (1:1million).

ext
Extension of the file(s) to be downloaded. Available formats are "geojson", "shp", "svg", "json", "gdb". See Details.

recursive
Tries to unzip recursively the zip files (if any) included in the initial bulk download (case of ext = "shp".

Source

https://gisco-services.ec.europa.eu/distribution/v2/

Value

Silent function.

Details

See the years available in gisco_get

The usual extension used across giscoR is "geojson", however other formats are already available on GISCO.

This function helps building a personal shape library on cache_dir (or options(gisco_cache_dir = "path/to/dir"), if set by the user).

Note

For downloading specific files use gisco_get functions.

Examples

if (FALSE) {

# Countries 2016 - It would take some time
gisco_bulk_download(id_giscoR = "countries", resolution = "60")
}