Dataframe with different country code schemes and world regions

A dataframe containing conversions between different country code schemes (Eurostat/ISO2 and 3) as well as geographic regions as provided by the World Bank and the UN (M49). This dataset is extracted from countrycode package.

Format

A data frame object with 249 rows and 12 variables:

CNTR_CODE: Eurostat code of each country

iso2c: ISO 3166-1 alpha-2 code of each country

ISO3_CODE: ISO 3166-1 alpha-3 code of each country

iso.name.en: ISO English short name

cldr.short.en: English short name as provided by the Unicode Common Locale Data Repository https://cldr.unicode.org/translation/displaynames/countryregion-territory-names

continent: As provided by the World Bank

un.region.code: Numeric region code UN (M49)

un.region.name: Region name UN (M49)

un.regionintermediate.code: Numeric intermediate Region code UN (M49)

un.regionintermediate.name: Intermediate Region name UN (M49)

un.regionsub.code: Numeric sub-region code UN (M49)

un.regionsub.name: Sub-Region name UN (M49)

eu: Logical indicating if the country belongs to the European Union as per February 2021.

Source

countrycode::codelist v1.2.0.

See also

Examples

data(gisco_countrycode)