Download Geospatial Country Data from GISCO

Downloads a simple feature (sf) object.

gisco_get_countries(
  resolution = "20",
  year = "2016",
  crs = "4326",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  spatialtype = "RG",
  country = NULL,
  region = NULL
)

Arguments

resolution
Resolution of the geospatial data. One of

"60" (1:60million),

"20" (1:20million)

"10" (1:10million)

"03" (1:3million) or

"01" (1:1million).

year
Release year. One of "2001", "2006", "2010", "2013", "2016" or "2020"

crs
projection of the map: 4-digit EPSG code. One of:

"4326" - WGS84

"3035" - ETRS89 / ETRS-LAEA

"3857" - Pseudo-Mercator

cache
a logical whether to do caching. Default is TRUE.

update_cache
a logical whether to update cache.

cache_dir
a path to a cache directory. The directory have to exist. The NULL (default) uses and creates /gisco directory in the temporary directory from tempdir. The directory can also be set with options(gisco_cache_dir = <path>).

spatialtype
Type of geometry to be returned:

RG: Regions - Multipolygon

LB: Labels - Point

BN: Boundaries - Multilines

COASTL: coastlines - Multilines

INLAND: inland boundaries - Multilines

country
Optional. A character vector of ISO-3 country codes. See Details

region
Optional. A character vector of UN M49 region codes. Possible values are "Africa", "Americas", "Asia", "Europe", "Oceania". See Details and gisco_countrycode

Source

GISCO Countries

Value

a sf object.

Details

country and region only available when applicable. Some spatialtype datasets (as Multilines data-types) may not have country-level identifies.

Note

COPYRIGHT NOTICE

When data downloaded from this page http://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units is used in any printed or electronic publication, in addition to any other provisions applicable to the whole Eurostat website, data source will have to be acknowledged in the legend of the map and in the introductory page of the publication with the following copyright notice:

EN: (C) EuroGeographics for the administrative boundaries

FR: (C) EuroGeographics pour les limites administratives

DE: (C) EuroGeographics bezuglich der Verwaltungsgrenzen

For publications in languages other than English, French or German, the translation of the copyright notice in the language of the publication shall be used.

If you intend to use the data commercially, please contact EuroGeographics for information regarding their licence agreements.

See also

Author

dieghernan, https://github.com/dieghernan/

Examples

library(sf)

# Some data are already available for speed up the process
opar <- par(no.readonly = TRUE)
par(mar = c(2, 0, 0, 0))
africa2016 <-  gisco_get_countries(region = "Africa")
angola_namibia <-  gisco_get_countries(country = c("AGO", "NAM"))

plot(st_geometry(africa2016), bg = "#C6ECFF", col = NA)
plot(st_geometry(gisco_countries_20M_2016),
     col = "#E0E0E0",
     add = TRUE)
plot(st_geometry(africa2016), col = "#F6E1B9", add = TRUE)
plot(st_geometry(angola_namibia), col = "#FEFEE9", add = TRUE)

mtext(gisco_attributions(), side = 1, cex = 0.8)

# Change crs and resolution

cntries2020 <-
  gisco_get_countries(year = "2020",
                      crs = "3035",
                      resolution = "60")
plot(st_geometry(cntries2020), bg = "#C6ECFF", col = "#E0E0E0")
mtext(gisco_attributions(), side = 1, cex = 0.8)
par(opar)

# Several geometry types
coastl <-
  gisco_get_countries(spatialtype = "COASTL", resolution = "60")

inland <-
  gisco_get_countries(spatialtype = "INLAND", resolution = "60")