Get locations of healthcare services in Europe.

The dataset contains information on main healthcare services considered to be 'hospitals' by Member States.

Usage

gisco_get_healthcare(
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  country = NULL
)

Source

https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/healthcare-services

Arguments

cache

A logical whether to do caching. Default is TRUE. See About caching.

update_cache

A logical whether to update cache. Default is FALSE. When set to TRUE it would force a fresh download of the source .geojson file.

cache_dir

A path to a cache directory. See About caching.

verbose

Logical, displays information. Useful for debugging, default is FALSE.

country

Optional. A character vector of country codes. It could be either a vector of country names, a vector of ISO3 country codes or a vector of Eurostat country codes. Mixed types (as c("Turkey","US","FRA")) would not work. See also countrycode::countrycode().

Value

A POINT object.

Details

Files are distributed on EPSG:4326. Metadata available on https://gisco-services.ec.europa.eu/pub/healthcare/metadata.pdf.

About caching

You can set your cache_dir with gisco_set_cache_dir().

Sometimes cached files may be corrupt. On that case, try re-downloading the data setting update_cache = TRUE.

If you experience any problem on download, try to download the corresponding .geojson file by any other method and save it on your cache_dir. Use the option verbose = TRUE for debugging the API query.

For a complete list of files available check gisco_db.

See also

Author

dieghernan, https://github.com/dieghernan/

Examples

# \donttest{

health_BEL <- gisco_get_healthcare(country = "Belgium")
health_BEL
#> Simple feature collection with 272 features and 25 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 2.668537 ymin: 49.55318 xmax: 6.127014 ymax: 51.35934
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>          id                               hospital_name
#> 267 BE_VH12             C.H. UNIVERSITAIRE ANDRE VESALE
#> 268 BE_1020 ALGEMEEN ZIEKENHUIS ST.-JAN BRUGGE-OOSTENDE
#> 269 BE_1040                    ALGEMEEN ZIEKENHUIS ZENO
#> 270 BE_1060               ALGEMEEN ZIEKENHUIS GROENINGE
#> 271 BE_1100                   ALGEMEEN ZIEKENHUIS DELTA
#> 272 BE_1120                  JAN YPERMAN ZIEKENHUIS VZW
#> 273 BE_1150                   ALGEMEEN ZIEKENHUIS DELTA
#> 274 BE_1180                 ZIEKENHUISNETWERK ANTWERPEN
#> 275 BE_1190                 ZIEKENHUISNETWERK ANTWERPEN
#> 276 BE_1220                                 A.Z. MONICA
#>                     site_name      lat     lon                     street
#> 267          LEONARD DE VINCI 50.36193  4.3642           RUE DE GOZEE 706
#> 268 ST. FRANCISCUS - XAVERIUS 51.21308 3.22552          SPAANSE LOSKAAI 1
#> 269              BLANKENBERGE   51.308  3.1247      DR. VERHAEGHESTRAAT 1
#> 270                  REEPKAAI 50.82972 3.26091                 REEPKAAI 4
#> 271            BRUGSESTEENWEG 50.95403 3.12352          BRUGSESTEENWEG 90
#> 272                 POPERINGE  50.8559 2.74437                OOSTLAAN 11
#> 273                  WESTLAAN 50.93807 3.11549               WESTLAAN 123
#> 274           ZNA STUIVENBERG 51.22214 4.43526 LANGE BEELDEKENSSTRAAT 267
#> 275           ZNA ST. ERASMUS 51.20811 4.44691         LT. LIPPENSLAAN 55
#> 276                 ANTWERPEN 51.20176 4.41025          HARMONIESTRAAT 68
#>     house_number postcode                city cc country emergency cap_beds
#> 267         <NA>     6110 MONTIGNY-LE-TILLEUL BE Belgium        no      173
#> 268         <NA>     8000              BRUGGE BE Belgium        no      139
#> 269         <NA>     8370        BLANKENBERGE BE Belgium       yes       64
#> 270         <NA>     8500            KORTRIJK BE Belgium        no      100
#> 271         <NA>     8800           ROESELARE BE Belgium        no      323
#> 272         <NA>     8970           POPERINGE BE Belgium        no        0
#> 273         <NA>     8800           ROESELARE BE Belgium        no      179
#> 274         <NA>     2060         ANTWERPEN-6 BE Belgium       yes      320
#> 275         <NA>     2140          BORGERHOUT BE Belgium        no       87
#> 276         <NA>     2018         ANTWERPEN-1 BE Belgium       yes      201
#>     cap_prac cap_rooms facility_type public_private list_specs          tel
#> 267     <NA>      <NA>          <NA>           <NA>       <NA> 071/92.38.73
#> 268     <NA>      <NA>          <NA>           <NA>       <NA> 050/47.04.70
#> 269     <NA>      <NA>          <NA>           <NA>       <NA> 050/53.40.00
#> 270     <NA>      <NA>          <NA>           <NA>       <NA> 056/63.63.63
#> 271     <NA>      <NA>          <NA>           <NA>       <NA> 051/23.61.11
#> 272     <NA>      <NA>          <NA>           <NA>       <NA> 057/35.35.35
#> 273     <NA>      <NA>          <NA>           <NA>       <NA> 051/23.81.11
#> 274     <NA>      <NA>          <NA>           <NA>       <NA> 03/217.71.11
#> 275     <NA>      <NA>          <NA>           <NA>       <NA> 03/270.80.11
#> 276     <NA>      <NA>          <NA>           <NA>       <NA> 03/240.20.20
#>     email  url ref_date   pub_date geo_qual comments                      geom
#> 267  <NA> <NA> 01/02/20 19/10/2023       -1     <NA> POINT (4.364199 50.36193)
#> 268  <NA> <NA> 01/02/20 19/10/2023       -1     <NA> POINT (3.225525 51.21308)
#> 269  <NA> <NA> 01/02/20 19/10/2023       -1     <NA>   POINT (3.124703 51.308)
#> 270  <NA> <NA> 01/02/20 19/10/2023       -1     <NA> POINT (3.260914 50.82972)
#> 271  <NA> <NA> 01/02/20 19/10/2023       -1     <NA> POINT (3.123518 50.95403)
#> 272  <NA> <NA> 01/02/20 19/10/2023       -1     <NA>  POINT (2.744373 50.8559)
#> 273  <NA> <NA> 01/02/20 19/10/2023       -1     <NA> POINT (3.115493 50.93807)
#> 274  <NA> <NA> 01/02/20 19/10/2023       -1     <NA> POINT (4.435258 51.22214)
#> 275  <NA> <NA> 01/02/20 19/10/2023       -1     <NA> POINT (4.446906 51.20811)
#> 276  <NA> <NA> 01/02/20 19/10/2023       -1     <NA> POINT (4.410254 51.20176)
# }