Get GISCO NUTS sf polygons, points and lines

Returns NUTS regions polygons, lines and points at a specified scale, as provided by GISCO.

NUTS are provided at three different levels:

"0": Country level

"1": Groups of states/regions

"2": States/regions

"3": Counties/provinces/districts

Note that NUTS-level definition may vary across countries. See also https://ec.europa.eu/eurostat/web/nuts/background.

Arguments

year

Release year of the file. One of "2003", "2006, "2010", "2013", "2016" or "2021".

epsg

projection of the map: 4-digit EPSG code. One of:

"4258": ETRS89

"4326": WGS84

"3035": ETRS89 / ETRS-LAEA

"3857": Pseudo-Mercator

cache

A logical whether to do caching. Default is TRUE. See About caching.

update_cache

A logical whether to update cache. Default is FALSE. When set to TRUE it would force a fresh download of the source .geojson file.

cache_dir

A path to a cache directory. See About caching.

verbose

Logical, displays information. Useful for debugging, default is FALSE.

resolution

Resolution of the geospatial data. One of

"60": 1:60million

"20": 1:20million

"10": 1:10million

"03": 1:3million

"01": 1:1million

spatialtype

Type of geometry to be returned:

"BN": Boundaries - LINESTRING object.

"LB": Labels - POINT object.

"RG": Regions - MULTIPOLYGON/POLYGON object.

country

Optional. A character vector of country codes. It could be either a vector of country names, a vector of ISO3 country codes or a vector of Eurostat country codes. Mixed types (as c("Turkey","US","FRA")) would not work. See also countrycode::countrycode().

nuts_id

Optional. A character vector of NUTS IDs.

nuts_level

NUTS level. One of "0", "1", "2" or "3". See Description.

Source

https://gisco-services.ec.europa.eu/distribution/v2/

Value

A sf object specified by spatialtype.

About caching

You can set your cache_dir with gisco_set_cache_dir().

Sometimes cached files may be corrupt. On that case, try re-downloading the data setting update_cache = TRUE.

If you experience any problem on download, try to download the corresponding .geojson file by any other method and save it on your cache_dir. Use the option verbose = TRUE for debugging the API query.

For a complete list of files available check gisco_db.

See also

Examples