Get geospatial units data from GISCO API

Download individual shapefiles of units. Unlike gisco_get_countries, gisco_get_nuts or gisco_get_urban_audit, that downloads a full dataset and applies filters, gisco_get_units downloads a single shapefiles for each unit.

gisco_get_units(
  id_giscoR = "nuts",
  unit = "ES4",
  mode = "sf",
  year = "2016",
  epsg = "4326",
  cache = TRUE,
  update_cache = FALSE,
  cache_dir = NULL,
  verbose = FALSE,
  resolution = "20",
  spatialtype = "RG"
)

Arguments

id_giscoR
Select the unit type to be downloaded. Accepted values are 'nuts','countries' or 'urban_audit'.

unit
Unit ID to be downloaded. See Details.

mode
Controls the output of the function. Possible values are "sf" or "df". See Value and Details.

year, epsg, cache, update_cache, cache_dir, verbose, resolution
See gisco_get.

spatialtype
Type of geometry to be returned: "RG" for POLYGON and "LB" for POINT.

Source

GISCO API

Value

A sf object on mode = "sf" or a dataframe on mode = "df".

Details

The function can return a dataframe on mode = "df" or a sf object on mode = "sf"

In order to see the available unit ids with the required combination of what,year, first run the function on "df" mode. Once that you get the data frame you can select the required ids on the unit parameter.

On mode = "df" the only relevant parameters are what, year.

Note

countries file would be renamed on your cache_dir to avoid naming conflicts with nuts datasets.

See also

Author

dieghernan, https://github.com/dieghernan/

Examples

if (FALSE) {
library(sf)

if (gisco_check_access()) {
  cities <- gisco_get_units(
    id_giscoR = "urban_audit",
    mode = "df",
    year = "2020"
  )
  VAL <- cities[grep("Valencia", cities$URAU_NAME), ]
  #'   Order from big to small
  VAL <- VAL[order(as.double(VAL$AREA_SQM), decreasing = TRUE), ]

  VAL.sf <- gisco_get_units(
    id_giscoR = "urban_audit",
    year = "2020",
    unit = VAL$URAU_CODE
  )
  # Provincia
  Provincia <-
    gisco_get_units(
      id_giscoR = "nuts",
      unit = c("ES523"),
      resolution = "01"
    )

  # Surrounding area
  NUTS1 <-
    gisco_get_units(
      id_giscoR = "nuts",
      unit = c("ES5"),
      resolution = "01"
    )

  # Plot
  plot(
    st_geometry(Provincia),
    col = "gray1",
    border = "grey50",
    lwd = 3
  )
  plot(st_geometry(NUTS1),
    border = "grey50",
    lwd = 3,
    add = TRUE
  )
  plot(
    st_geometry(VAL.sf),
    col = c("deeppink4", "brown2", "khaki1"),
    add = TRUE
  )
  box()
  title(
    "Urban Audit - Valencia (ES)",
    sub = gisco_attributions("es"),
    line = 1,
    cex.sub = 0.7
  )
}
}